+++ src/main/java/org/recordy/server/external/AwsConfig.java
@@ -1,4 +1,4 @@
-package org.recordy.server.common.config;
+package org.recordy.server.external;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
